examify/backend/requirements.txt
Flask==3.0.3
Flask-CORS==4.0.1
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
openai==1.30.1
PyPDF2==3.0.1
python-pptx==0.6.23
python-magic==0.4.27
Werkzeug==3.0.3
SQLAlchemy==2.0.30
gunicorn==22.0.0
